chore(config): refactor package structure for AppConfig and EnvironmentConfig
